web_ui/requirements.txt
nicegui
pyserial
picamera2
